MessageList add header

+++ Android/messagelist/src/main/java/cn/jiguang/imui/utils/DisplayUtil.java
@@ -5,7 +5,7 @@ import android.util.TypedValue;
 
 
 public class DisplayUtil {
-    public static int dp2px(Context context, int dp) {
+    public static int dp2px(Context context, float dp) {
         return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp,
                 context.getResources().getDisplayMetrics());
     }
